+++ healthcheck_test.go
@@ -7,7 +7,9 @@ import (
 	"fmt"
 	"net/http"
 	"net/http/httptest"
+	"sync"
 	"testing"
+	"time"
 )
 
 func TestHealthChecks_Default(t *testing.T) {
@@ -27,6 +29,74 @@ func TestHealthChecks_Default(t *testing.T) {
 }
 
 func TestHealthChecks_Custom(t *testing.T) {
+	t.Run("with multiple tests", func(t *testing.T) {
+		defer resetTests()
+
+		executedTests := 0
+		var l sync.Mutex
+		tstFuncFail := func(_ context.Context) (Status, error) {
+			l.Lock()
+			defer l.Unlock()
+			executedTests++
+			return Unavailable, errors.New("unavailable")
+		}
+		tstFuncSuccess := func(_ context.Context) (Status, error) {
+			l.Lock()
+			defer l.Unlock()
+			executedTests++
+			return Available, nil
+		}
+		RegisterTest("custom-failure1", tstFuncFail)
+		RegisterTest("custom-success1", tstFuncSuccess)
+		RegisterTest("custom-failure2", tstFuncFail)
+		RegisterTest("custom-success2", tstFuncSuccess)
+
+		hc, _, err := getHealth()
+		if err != nil {
+			t.Fatalf("Expected no error, got '%s'", err.Error())
+		}
+
+		// account for default test
+		if ln := len(hc.Tests); ln != 5 {
+			t.Fatalf("Expected '%d' tests, got '%d'", 5, ln)
+		}
+
+		if executedTests != 4 {
+			t.Fatalf("Expected '%d' tests to be executed, got '%d'", 4, executedTests)
+		}
+	})
+
+	t.Run("with an exceeded deadline", func(t *testing.T) {
+		defer resetTests()
+
+		Timeout = 50 * time.Millisecond
+		tstFunc := func(_ context.Context) (Status, error) {
+			time.Sleep(time.Second)
+			return Available, nil
+		}
+		RegisterTest("success", tstFunc)
+
+		hc, _, err := getHealth()
+		if err != nil {
+			t.Fatalf("Expected no error, got '%s'", err.Error())
+		}
+
+		// our registered test shouldn't be skipped
+		if ln := len(hc.Tests); ln != 2 {
+			t.Fatalf("Expected '%d' tests, got '%d'", 1, ln)
+		}
+
+		// our registered test should be marked unavailable despite what the
+		// outcome is as the response will be too late
+		if hc.Tests["success"].Status != Unavailable {
+			t.Fatalf("Expected 'success' test to be Unavailable, got '%s'", hc.Tests["success"].Status)
+		}
+
+		if hc.Tests["success"].Error != ErrTimeout {
+			t.Fatalf("Expected 'success' test to be timeout")
+		}
+	})
+
 	t.Run("with a failing test", func(t *testing.T) {
 		defer resetTests()
 
@@ -165,5 +235,6 @@ func getHealth() (HealthCheck, int, error) {
 
 func resetTests() {
 	healthCheckTests = map[string]TestFunc{}
+	Timeout = 5 * time.Second
 	RegisterTest("default", defaultCheck)
 }